Fixes #1666: Remove duplicate entry of Sieve of Eratosthenes
