+++ Maths/test/SieveOfEratosthenes.test.js
@@ -1,14 +1,12 @@
 import { sieveOfEratosthenes } from '../SieveOfEratosthenes'
 import { PrimeCheck } from '../PrimeCheck'
 
-describe('should return an array of prime booleans', () => {
-  it('should have each element in the array as a prime boolean', () => {
-    const n = 30
+describe('should return an array of prime numbers', () => {
+  it('should have each element in the array as a prime numbers', () => {
+    const n = 100
     const primes = sieveOfEratosthenes(n)
-    primes.forEach((primeBool, index) => {
-      if (primeBool) {
-        expect(PrimeCheck(index)).toBeTruthy()
-      }
+    primes.forEach((prime) => {
+      expect(PrimeCheck(prime)).toBeTruthy()
     })
   })
 })
